Program to compute the Calculator Performance Index
At https://thimet.de/CalcCollection/CalcPerformance.html there is a nice overview of the computing speed of various calculators. The "Ur-Boris", or rather one of his descendants (Elektronika MK-61) is at the top of the list (he needs the longest for the example program). Since the sample program is quite simple, it was quickly adapted for my boris4 and the Boris Voyager.
Boris4 with 1MHz clock comes with 26 seconds to an index of 13 and is on par with the HP-42S. Boris Voyager comes with 10.2seconds to an index of 33 in about the HP-32S. I'm be a little proud of that.

Enclosed the programs:
bench4.b4s - source code for boris4
bench4.bor - binary program for boris4, can be loaded using either BC or terminal
PROG_3.B5M - source code / loadable program for boris voyager (2be loaded from SD card)